# unify c2 and TH allocator