--- a/caffe2/core/context_test.cc
+++ b/caffe2/core/context_test.cc
@@ -9,7 +9,7 @@ namespace caffe2 {
 TEST(CPUContextTest, TestAllocAlignment) {
   for (int i = 1; i < 10; ++i) {
     auto data = CPUContext::New(i);
-    EXPECT_EQ((reinterpret_cast<size_t>(data.get()) % gCaffe2Alignment), 0);
+    EXPECT_EQ((reinterpret_cast<size_t>(data.get()) % gAlignment), 0);
     // data is freed when out of scope
   }
 }