@SsnL 2019-02-08
Things got slightly slower overall. @zou3519 will be sad. 

Why is `BM_MakeVariableFromTensor` much slower?